Bundle standard contracts with MIRAI binary
