+++ checker/tests/run-pass/vec_pop.rs
@@ -4,9 +4,7 @@
 // LICENSE file in the root directory of this source tree.
 //
 
-// A test that exercises visit_used_move with a structured value.
-
-#![feature(type_alias_enum_variants)]
+// A test that uses built-in contracts for the Vec struct.
 
 #[macro_use]
 extern crate mirai_annotations;
@@ -21,56 +19,4 @@ pub fn main() {
     verify!(v.len() == old_len + 1);
     v.pop();
     verify!(v.len() == old_len);
-}
-
-pub mod foreign_contracts {
-    pub mod core {
-        pub mod usize {
-            pub const MAX: usize = 4294967295;
-        }
-    }
-
-    pub mod alloc {
-        pub mod vec {
-            pub struct Vec<T> {
-                _phantom: std::marker::PhantomData<T>,
-                pub len: usize,
-            }
-            
-            impl<T> Vec<T> {
-                pub fn new() -> Vec<T> {
-                    let res = Vec {
-                        _phantom: std::marker::PhantomData,
-                        len: 0,
-                    };
-                    verify!(res.len == 0);
-                    res
-                }
-
-                pub fn len(&self) -> usize {
-                    let old_len = self.len;
-                    assume!(self.len == old_len);
-                    self.len
-                }
-
-                pub fn push(&mut self, _value: T) {
-                    precondition!(self.len < std::usize::MAX);
-                    let old_len = self.len;
-                    self.len += 1;
-                    verify!(self.len == old_len + 1);
-                }
-
-                pub fn pop(&mut self) -> Option<T> {
-                    if self.len == 0 {
-                        None
-                    } else {
-                        let old_len = self.len;
-                        self.len -= 1;
-                        verify!(self.len == old_len - 1);
-                        result!()
-                    }
-                }
-            }
-        }
-    }
 }
