+++ standard_contracts/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+
+name = "mirai-standard-contracts"
+version = "0.0.1"
+authors = ["Herman Venter <hermanv@fb.com>"]
+description = "Specifications of standard library functions for MIRAI"
+repository = "https://github.com/facebookexperimental/MIRAI"
+readme = "README.md"
+license = "MIT"
+edition = "2018"
+
+[lib]
+test = false # we have no unit tests
+doctest = false # and no doc tests
+
+[dependencies]
+mirai-annotations = { path = "../annotations" }
